+++ python/cohorte/boot/boot.py
@@ -61,7 +61,7 @@ import pelix.framework
 __docformat__ = "restructuredtext en"
 
 # Boot module version
-__version_info__ = (1, 0, 1)
+__version_info__ = (1, 1, 0)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # ------------------------------------------------------------------------------