+++ python/cohorte/boot/looper/qt.py
@@ -44,7 +44,7 @@ import PyQt4.QtGui as QtGui
 # ------------------------------------------------------------------------------
 
 # Module version
-__version_info__ = (1, 0, 1)
+__version_info__ = (1, 1, 0)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format