+++ python/cohorte/forker/starters/exe.py
@@ -5,7 +5,7 @@ Starts basic executable file as isolates
 
 :author: Thomas Calmant
 :license: Apache Software License 2.0
-:version: 1.0.1
+:version: 1.1.0
 
 ..
 
@@ -42,7 +42,7 @@ import cohorte.forker.starters.common as common
 __docformat__ = "restructuredtext en"
 
 # Version
-__version_info__ = (1, 0, 1)
+__version_info__ = (1, 1, 0)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # ------------------------------------------------------------------------------