+++ python/cohorte/local/discovery.py
@@ -31,7 +31,7 @@ belonging to the same node and contacts them respectively.
 """
 
 # Module version
-__version_info__ = (0, 0, 5)
+__version_info__ = (1, 1, 0)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format