+++ python/cohorte/monitor/basic.py
@@ -5,7 +5,7 @@ COHORTE Monitor: Monitor included in the F/M/N process
 
 :author: Thomas Calmant
 :license: Apache Software License 2.0
-:version: 1.0.1
+:version: 1.1.0
 
 ..
 
@@ -56,7 +56,7 @@ from herald.beans import Message
 __docformat__ = "restructuredtext en"
 
 # Version
-__version_info__ = (1, 0, 1)
+__version_info__ = (1, 1, 0)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # ------------------------------------------------------------------------------