+++ python/cohorte/monitor/fsm.py
@@ -32,7 +32,7 @@ from cohorte.utils.statemachine import StateMachine
 __docformat__ = "restructuredtext en"
 
 # Version
-__version_info__ = (1, 0, 1)
+__version_info__ = (1, 1, 0)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # ------------------------------------------------------------------------------