+++ python/cohorte/monitor/status.py
@@ -40,7 +40,7 @@ import cohorte.monitor.fsm as fsm
 __docformat__ = "restructuredtext en"
 
 # Version
-__version_info__ = (1, 0, 1)
+__version_info__ = (1, 1, 0)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # ------------------------------------------------------------------------------