+++ python/cohorte/repositories/beans.py
@@ -46,7 +46,7 @@ from pelix.utilities import is_string
 __docformat__ = "restructuredtext en"
 
 # Version
-__version_info__ = (1, 0, 1)
+__version_info__ = (1, 1, 0)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # ------------------------------------------------------------------------------