+++ python/cohorte/shell/agent.py
@@ -48,7 +48,7 @@ import herald.beans as beans
 # ------------------------------------------------------------------------------
 
 # Version
-__version_info__ = (1, 0, 1)
+__version_info__ = (1, 1, 0)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format