+++ python/cohorte/utils/posix.py
@@ -42,7 +42,7 @@ import cohorte.utils as utils
 __docformat__ = "restructuredtext en"
 
 # Version
-__version_info__ = (1, 0, 1)
+__version_info__ = (1, 1, 0)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # ------------------------------------------------------------------------------