+++ python/cohorte/utils/ruleengine.py
@@ -39,7 +39,7 @@ import intellect.Intellect
 __docformat__ = "restructuredtext en"
 
 # Version
-__version_info__ = (1, 0, 1)
+__version_info__ = (1, 1, 0)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # ------------------------------------------------------------------------------