+++ CHANGELOG.rst
@@ -1,11 +1,19 @@
-1.0.4 (current dev)
------
+Changes form the 1.0.3 to 1.1.0
+-------------------------------
+
+** New Features
+    * Local Discovery 
+
+** Improvements
+    * Adding Forker's HTTP Port to OSGi java framework properties
+    * Using Cohorte-Utilities 1.0.6
 
 ** Bug Fix
+    * Java isolates are not started if no vm_args configuration was given [isandlatech/cohorte-platforms#79]
     * Normalize paths coming from PYTHONPATH #34
 
-1.0.3
------
+Changes form the 1.0.2 to 1.0.3
+-------------------------------
 
 ** Bug Fix
     * OSGi framework user-provided extra packages are not considered [isandlaTech/cohorte-runtime#36]